requirements.txt
numpy==1.21.6
pandas==1.3.5
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
